organized functions on article page into several files, updated some code to handle errors and to be imported/exported
